Update LineBox.java
Updated method `checkPagePosition` so that if a page break is required but the element's parent has a CSS overflow attribute set to hidden (and its height is on auto) then the elements size is adjusted so as to not overflow the page and the page break is cancelled. The result will be a clipped / cropped element at the bottom of the page instead of the element being moved over onto the next page.